Init Ambient and Directional lights (and other minor changes)
 The first light we add is an ambient light, meaning this light will radiate from all directions, which is useful if we just want our entire scene to be lit up without doing anything complex. We also pass one argument to our light, a hexidecimal color code, prefixed in three.js by "0x". Here, we have given our ambient light a grey tint. After we create the light, we must then add it to the scene.

The second light we create is a directional light. This is a light that will emanate from a certain position, which we set to be (20, 10, 20), the right-upper corner on the near side. The two arguments we pass here are the hex color, and the intensity. We give it a white-yellow tint, and the maximum intensity of 1.